// tech-blog-ui/src/app/api/admin-api.service.generated.ts
export class PostCategoryDto implements IPostCategoryDto {
    id?: string;
    name?: string | undefined;
    slug?: string | undefined;
    parentId?: string | undefined;
    isActive?: boolean;
    dateCreated?: Date;
    dateModified?: Date | undefined;
    seoDescription?: string | undefined;
    sortOrder?: number;

    constructor(data?: IPostCategoryDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.slug = _data["slug"];
            this.parentId = _data["parentId"];
            this.isActive = _data["isActive"];
            this.dateCreated = _data["dateCreated"] ? new Date(_data["dateCreated"].toString()) : <any>undefined;
            this.dateModified = _data["dateModified"] ? new Date(_data["dateModified"].toString()) : <any>undefined;
            this.seoDescription = _data["seoDescription"];
            this.sortOrder = _data["sortOrder"];
        }
    }

    static fromJS(data: any): PostCategoryDto {
        data = typeof data === 'object' ? data : {};
        let result = new PostCategoryDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["slug"] = this.slug;
        data["parentId"] = this.parentId;
        data["isActive"] = this.isActive;
        data["dateCreated"] = this.dateCreated ? this.dateCreated.toISOString() : <any>undefined;
        data["dateModified"] = this.dateModified ? this.dateModified.toISOString() : <any>undefined;
        data["seoDescription"] = this.seoDescription;
        data["sortOrder"] = this.sortOrder;
        return data;
    }
}

export interface IPostCategoryDto {
    id?: string;
    name?: string | undefined;
    slug?: string | undefined;
    parentId?: string | undefined;
    isActive?: boolean;
    dateCreated?: Date;
    dateModified?: Date | undefined;
    seoDescription?: string | undefined;
    sortOrder?: number;
}
// ...
export class PostDto implements IPostDto {
    id?: string;
    name?: string | undefined;
    slug?: string | undefined;
    description?: string | undefined;
    thumbnail?: string | undefined;
    viewCount?: number;
    dateCreated?: Date;
    isPaid?: boolean;
    royaltyAmount?: number;
    paidDate?: Date | undefined;
    categoryId?: string;
    content?: string | undefined;
    authorUserId?: string;
    source?: string | undefined;
    tags?: string | undefined;
    seoDescription?: string | undefined;
    dateModified?: Date | undefined;
    status?: PostStatus;

    constructor(data?: IPostDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.slug = _data["slug"];
            this.description = _data["description"];
            this.thumbnail = _data["thumbnail"];
            this.viewCount = _data["viewCount"];
            this.dateCreated = _data["dateCreated"] ? new Date(_data["dateCreated"].toString()) : <any>undefined;
            this.isPaid = _data["isPaid"];
            this.royaltyAmount = _data["royaltyAmount"];
            this.paidDate = _data["paidDate"] ? new Date(_data["paidDate"].toString()) : <any>undefined;
            this.categoryId = _data["categoryId"];
            this.content = _data["content"];
            this.authorUserId = _data["authorUserId"];
            this.source = _data["source"];
            this.tags = _data["tags"];
            this.seoDescription = _data["seoDescription"];
            this.dateModified = _data["dateModified"] ? new Date(_data["dateModified"].toString()) : <any>undefined;
            this.status = _data["status"];
        }
    }

    static fromJS(data: any): PostDto {
        data = typeof data === 'object' ? data : {};
        let result = new PostDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["slug"] = this.slug;
        data["description"] = this.description;
        data["thumbnail"] = this.thumbnail;
        data["viewCount"] = this.viewCount;
        data["dateCreated"] = this.dateCreated ? this.dateCreated.toISOString() : <any>undefined;
        data["isPaid"] = this.isPaid;
        data["royaltyAmount"] = this.royaltyAmount;
        data["paidDate"] = this.paidDate ? this.paidDate.toISOString() : <any>undefined;
        data["categoryId"] = this.categoryId;
        data["content"] = this.content;
        data["authorUserId"] = this.authorUserId;
        data["source"] = this.source;
        data["tags"] = this.tags;
        data["seoDescription"] = this.seoDescription;
        data["dateModified"] = this.dateModified ? this.dateModified.toISOString() : <any>undefined;
        data["status"] = this.status;
        return data;
    }
}

export interface IPostDto {
    id?: string;
    name?: string | undefined;
    slug?: string | undefined;
    description?: string | undefined;
    thumbnail?: string | undefined;
    viewCount?: number;
    dateCreated?: Date;
    isPaid?: boolean;
    royaltyAmount?: number;
    paidDate?: Date | undefined;
    categoryId?: string;
    content?: string | undefined;
    authorUserId?: string;
    source?: string | undefined;
    tags?: string | undefined;
    seoDescription?: string | undefined;
    dateModified?: Date | undefined;
    status?: PostStatus;
}
// ...
export enum PostStatus {
    _0 = 0,
    _1 = 1,
    _2 = 2,
    _3 = 3,
}